refactor(ts): migrate geoSearch
Did not migrate the renderer, since that's not really of value to customers

DX-1594
