+++ src/widgets/geo-search/defaultTemplates.ts
